Optimization axios queries in components. Minor edits
