+++ src/components/tables/DefaultTable.vue
@@ -1,58 +1,56 @@
 <template>
-  <div v-if="!loading">
-    <transition appear enter-active-class="animated fadeIn" leave-active-class="animated fadeOut">
-      <q-table
-        :title="title"
-        :rows="rows"
-        :columns="props.columns"
-        :pagination="tablePagination"
-        row-key="title"
-        dense
-        flat
-        class="q-pa-sm custom-rounded text-blue-grey-9 shadow-sm"
-      >
-        <template v-slot:top-right>
-          <q-btn
-            @click="addModal = !addModal"
-            size="16px"
-            color="primary"
-            icon="add"
-            class="custom-rounded"
-          ></q-btn>
-        </template>
+  <div>
+    <q-table
+      :title="title"
+      :rows="rows"
+      :columns="props.columns"
+      :pagination="tablePagination"
+      row-key="title"
+      dense
+      flat
+      class="q-pa-sm custom-rounded text-blue-grey-9 shadow-sm"
+    >
+      <template v-slot:top-right>
+        <q-btn
+          @click="addModal = !addModal"
+          size="16px"
+          color="primary"
+          icon="add"
+          class="custom-rounded"
+        ></q-btn>
+      </template>
 
-        <template v-slot:body="props">
-          <q-tr :props="props">
-            <q-td key="id" :props="props">
-              {{ props.row.id }}
-            </q-td>
-            <q-td key="title" :props="props">
-              <q-badge color="blue-grey-10 text-subtitle2 shadow-1 custom-rounded">
-                <q-input
-                  :model-value="props.row.title"
-                  @update:model-value="(val) => onUpdateTitle(val, props.row)"
-                  @blur="updateItem(path, props.row)"
-                  dense
-                  borderless
-                  input-style="text-align: center; color: white"
-                >
-                </q-input>
-              </q-badge>
-            </q-td>
-            <q-td key="destroy" :props="props">
-              <q-btn
-                @click="((itemToRemove = props.row), (modalRemove = true))"
-                icon="close"
-                size="12px"
-                color="negative"
-                round
+      <template v-slot:body="props">
+        <q-tr :props="props">
+          <q-td key="id" :props="props">
+            {{ props.row.id }}
+          </q-td>
+          <q-td key="title" :props="props">
+            <q-badge color="blue-grey-10 text-subtitle2 shadow-1 custom-rounded">
+              <q-input
+                :model-value="props.row.title"
+                @update:model-value="(val) => onUpdateTitle(val, props.row)"
+                @blur="updateItem(path, props.row)"
                 dense
-              ></q-btn>
-            </q-td>
-          </q-tr>
-        </template>
-      </q-table>
-    </transition>
+                borderless
+                input-style="text-align: center; color: white"
+              >
+              </q-input>
+            </q-badge>
+          </q-td>
+          <q-td key="destroy" :props="props">
+            <q-btn
+              @click="((itemToRemove = props.row), (modalRemove = true))"
+              icon="close"
+              size="12px"
+              color="negative"
+              round
+              dense
+            ></q-btn>
+          </q-td>
+        </q-tr>
+      </template>
+    </q-table>
     <q-dialog
       v-model="modalRemove"
       persistent
@@ -68,18 +66,18 @@
       <default-form @storeItem="(item) => postItem(path, item)" />
     </q-dialog>
   </div>
-  <InnderLoading v-else />
 </template>
 <script setup>
-import { ref, onMounted } from 'vue'
+import { ref, onMounted, defineEmits } from 'vue'
 import ConfirmationCard from '../ui/ConfirmationCard.vue'
 import DefaultForm from 'src/components/forms/DefaultForm.vue'
-import InnderLoading from '../ui/InnderLoading.vue'
 import { getData } from 'src/utils/http/get'
 import { deleteData } from 'src/utils/http/delete'
 import { postData } from 'src/utils/http/post'
 import { patchData } from 'src/utils/http/patch'
 import notify from 'src/plugins/notify'
+
+const emit = defineEmits(['success'])
 
 const props = defineProps({
   title: {
@@ -97,8 +95,6 @@ const props = defineProps({
 
 const path = props.path
 
-const loading = ref(true)
-
 const addModal = ref(false)
 const modalRemove = ref(false)
 const itemToRemove = ref({})
@@ -109,7 +105,7 @@ const getItems = async (path) => {
   try {
     const response = await getData(path)
     rows.value = response
-    loading.value = false
+    emit('success')
   } catch (e) {
     console.error(e)
   }
@@ -123,6 +119,7 @@ const postItem = async (path, item) => {
     notify.success('Элемент добавлен')
   } catch (e) {
     console.error(e)
+    notify.error('Ошибка при добавлении')
   } finally {
     loading()
   }
@@ -137,6 +134,7 @@ const deleteItem = async (path, id) => {
     notify.success('Элемент удален')
   } catch (e) {
     console.error(e)
+    notify.error('Ошибка при удалении')
   } finally {
     loading()
   }
@@ -155,6 +153,7 @@ const updateItem = async (path, row) => {
     notify.success('Данные обновлены')
   } catch (e) {
     console.error(e)
+    notify.error('Ошибка при изменении')
     getData(path)
   } finally {
     loading()
@@ -173,7 +172,6 @@ const updateLocalData = (data) => {
 
 const removeLocalData = (id) => {
   const index = rows.value.findIndex((el) => el.id === id)
-  console.log(index, console.log(rows.value[index]))
   rows.value.splice(index, 1)
 }
