+++ src/css/style.scss
@@ -1,3 +1,8 @@
+a {
+  text-decoration: none;
+  color: inherit;
+}
+
 .gap-sm {
   gap: 10px;
 }
@@ -42,7 +47,6 @@
   box-shadow: 0 0px 2px 0 rgba(0, 0, 0, 0.164) !important;
 }
 
-a {
-  text-decoration: none;
-  color: inherit;
+.page-height {
+  min-height: calc(100vh - 66px);
 }
