+++ src/pages/shows/ProductShowPage.vue
@@ -1,6 +1,6 @@
 <template>
-  <div class="q-pa-md">
-    <q-card class="custom-rounded shadow-sm q-pa-sm" flat>
+  <div class="q-pa-md relative-position page-height">
+    <q-card v-show="!loading" class="custom-rounded shadow-sm q-pa-sm" flat>
       <q-toolbar>
         <q-toolbar-title>{{ tab === 'product' ? 'Продукт' : 'Ключи' }}</q-toolbar-title>
         <q-tabs
@@ -18,22 +18,31 @@
       <q-separator></q-separator>
       <q-tab-panels v-model="tab" animated keep-alive>
         <q-tab-panel name="product">
-          <ProductShow @getProductId="(productId) => (productId = productId)" />
+          <ProductShow
+            @getProductId="(productId) => (productId = productId)"
+            @success="loading = false"
+          />
         </q-tab-panel>
-        <q-tab-panel name="keys"> <ActivationKeysTable :productId="productId" /> </q-tab-panel>
+        <q-tab-panel name="keys">
+          <ActivationKeysTable :productId="+productId" />
+        </q-tab-panel>
       </q-tab-panels>
     </q-card>
+    <InnerLoading :loading="loading" size="200px" />
   </div>
 </template>
 <script setup>
 import { ref } from 'vue'
-import ProductShow from 'src/components/blocks/ProductShow.vue'
+import ProductShow from 'src/components/shows/ProductShow.vue'
 import ActivationKeysTable from 'src/components/tables/ActivationKeysTable.vue'
 import { useRoute } from 'vue-router'
+import InnerLoading from 'src/components/ui/InnerLoading.vue'
 
 const route = useRoute()
 
 const tab = ref('product')
 const productId = route.params.productId
+
+const loading = ref(!productId ? false : true)
 </script>
 <style lang=""></style>
